testing a new releases distro

+++ install.sh
@@ -11,7 +11,7 @@ if [ "${ID}" = "fedora" ]; then
 elif [ "${ID}" = "ubuntu" ] ; then
  package_manager="apt"
 
-elif [ "${ID}" = "debian" ] ; then
+else
   echo "NOK"
 
 fi
@@ -31,6 +31,10 @@ function install_basics {
   sudo $package_manager install curl flatpak yakuake openssh-server xterm zenity solaar \
                         git vim htop most zsh bat git-extras -y
   sudo dconf update
+}
+
+function add_flathub {
+  sudo flatpak remote-add --if-not-exists flathub https://flathub.org/repo/flathub.flatpakrepo
 }
 
 function flatpak_packages {
@@ -62,65 +66,58 @@ function install_fonts {
 function first_run {
   updated
   install_basics
+  add_flathub
   flatpak_packages
   install_fonts
-#  dont_need_this
-#  sysctl_set
-#  ssh_set
-
-#  repos_set
 }
 
 first_run
 
 
-# function dont_need_this {
-#     sudo su - root -c 'cat <<EOT >>/etc/fstab
-# tmpfs /tmp tmpfs defaults,noatime,mode=1777 0 0
-# tmpfs /var/tmp tmpfs defaults,noatime,mode=1777 0 0
-# tmpfs /var/log tmpfs defaults,noatime,mode=0755 0 0
-# }
 
-function sysctl_set {
-    sudo su - root -c 'curl https://raw.githubusercontent.com/Esl1h/UAI/main/conf/sysctl.conf >>/etc/sysctl.conf'
-    sudo sysctl -p
-}
-
-function ssh_set {
-  sudo su - root -c 'curl https://raw.githubusercontent.com/Esl1h/UAI/main/conf/ssh_config >/etc/ssh/ssh_config'
-  sudo systemctl enable ssh
-  sudo systemctl start ssh
-}
 
 function repos_set {
   # Softmaker Office
-  wget -qO - https://shop.softmaker.com/repo/linux-repo-public.key | sudo apt-key add - && \
-  sudo echo "deb https://shop.softmaker.com/repo/apt stable non-free" | sudo tee  /etc/apt/sources.list.d/softmaker.list
-  # Flathub
-  sudo flatpak remote-add --if-not-exists flathub https://flathub.org/repo/flathub.flatpakrepo
+  wget -qO - https://shop.softmaker.com/repo/linux-repo-public.key | sudo apt-key add -
   # NextDNS
   sudo wget -qO /usr/share/keyrings/nextdns.gpg https://repo.nextdns.io/nextdns.gpg
-  echo "deb [signed-by=/usr/share/keyrings/nextdns.gpg] https://repo.nextdns.io/deb stable main" | sudo tee /etc/apt/sources.list.d/nextdns.list
+}
 
+function install_softmaker-office-nx {
+  if [ "${ID}" = "fedora" ]; then
+      sudo -E dnf update -y
+      sudo -E dnf install softmaker-office-nx -y
+
+elif [ "${ID}" = "ubuntu" ] ; then
+      sudo echo "deb https://shop.softmaker.com/repo/apt stable non-free" | sudo tee  /etc/apt/sources.list.d/softmaker.list
+      sudo apt update
+      sudo apt install softmaker-office-nx
+else
+  echo "NOK"
+
+fi
 }
 
 
+function install_nextdns {
+  if [ "${ID}" = "fedora" ]; then
+      sh -c "$(curl -sL https://nextdns.io/install)"
 
-function second_install {
-  sudo apt install nextdns
-  sudo apt install softmaker-office-nx -y
+elif [ "${ID}" = "ubuntu" ] ; then
+      echo "deb [signed-by=/usr/share/keyrings/nextdns.gpg] https://repo.nextdns.io/deb stable main" | sudo tee /etc/apt/sources.list.d/nextdns.list
+      read -p "enter your NextDNS ID: " nextdnsid
+      sudo nextdns install -config $nextdnsid -report-client-info -auto-activate
+else
+  echo "NOK"
+
+fi
 }
 
+repos_set
+updated
+install_softmaker-office-nx
+install_nextdns
 
-
-function nextdns_set {
-  sudo nextdns install -config $nextdnsid -report-client-info -auto-activate
-}
-
-# first_run
-# updated
-# #nextdns_set CHANGE CONFIG ID!
-# second_install
 
 
 function install_zsh_ohmyzsh {
@@ -135,22 +132,41 @@ function install_zsh_ohmyzsh {
       mkdir -p ~/.oh-my-zsh/completions
       git clone https://github.com/zsh-users/zsh-syntax-highlighting.git  ~/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting
       git clone https://github.com/zsh-users/zsh-autosuggestions          ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions
-      git clone --depth 1 https://github.com/junegunn/fzf.git             ~/.fzf
-      ~/.fzf/install
+      # git clone --depth 1 https://github.com/junegunn/fzf.git             ~/.fzf
+      # ~/.fzf/install
 
       # powerlevel10k zsh theme
       git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ~/.oh-my-zsh/custom/themes/powerlevel10k
 }
 
-function config_zsh_ohmyzsh {
-    # config zshrc file:
-    rm ~/.zshrc
-    wget -c https://raw.githubusercontent.com/Esl1h/UAI/main/conf/zshrc -O ~/.zshrc
-    echo export ZSH=\""$HOME"/.oh-my-zsh\" >>~/.zshrc
-    echo "source \$ZSH/oh-my-zsh.sh" >>~/.zshrc
+# function config_zsh_ohmyzsh {
+#     # config zshrc file:
+#     rm ~/.zshrc
+#     wget -c https://raw.githubusercontent.com/Esl1h/UAI/main/conf/zshrc -O ~/.zshrc
+#     echo export ZSH=\""$HOME"/.oh-my-zsh\" >>~/.zshrc
+#     echo "source \$ZSH/oh-my-zsh.sh" >>~/.zshrc
 
-    #wget -c https://raw.githubusercontent.com/Esl1h/UAI/main/conf/p10k.zsh -O ~/.p10k.zsh
-}
+#     #wget -c https://raw.githubusercontent.com/Esl1h/UAI/main/conf/p10k.zsh -O ~/.p10k.zsh
+# }
 
 # install_zsh_ohmyzsh
 # config_zsh_ohmyzsh
+
+
+# function dont_need_this {
+#     sudo su - root -c 'cat <<EOT >>/etc/fstab
+# tmpfs /tmp tmpfs defaults,noatime,mode=1777 0 0
+# tmpfs /var/tmp tmpfs defaults,noatime,mode=1777 0 0
+# tmpfs /var/log tmpfs defaults,noatime,mode=0755 0 0
+# }
+
+# function sysctl_set {
+#     sudo su - root -c 'curl https://raw.githubusercontent.com/Esl1h/UAI/main/conf/sysctl.conf >>/etc/sysctl.conf'
+#     sudo sysctl -p
+# }
+
+# function ssh_set {
+#   sudo su - root -c 'curl https://raw.githubusercontent.com/Esl1h/UAI/main/conf/ssh_config >/etc/ssh/ssh_config'
+#   sudo systemctl enable ssh
+#   sudo systemctl start ssh
+# }
